feat(git): Introduce git live tracker from other remote than github
The code is straighforward and in bash(well... cuz it's faster and does
not require a third stuff to be installed on any UNIX distro).
- Check for the remote origin (if github, exit gracefully with 'return')
- extract informations related to the commit, repo, date
- files are tracked based on month-year
- then push

NOTE: the '/bin/echo' is to not call the 'echo' from bash built-in since
it may not consider some args passed like "-ne"

PS: will make it smarter later on, it's working fine for now.